[Rust][FFI] Move header helpers to common.rs
Move zerobus_alloc_header_array, zerobus_alloc_cstring, and
zerobus_free_headers from arrow.rs into common.rs, next to the
CHeader/CHeaders types they serve. No behavior, signature, or ABI
change; declarations move ahead of the Arrow functions in the
generated zerobus.h.
