+++ rust/ffi/NEXT_CHANGELOG.md
@@ -15,6 +15,8 @@
 
 ### Internal Changes
 
+- Move the header-callback helpers `zerobus_alloc_header_array`, `zerobus_alloc_cstring`, and `zerobus_free_headers` from `arrow.rs` into the shared `common.rs` module, next to the `CHeader`/`CHeaders` types they serve. No behavior, signature, or ABI change; their declarations move ahead of the Arrow functions in the generated `zerobus.h`.
+
 ### Behavior Changes
 
 ### Breaking Changes
