+++ rust/ffi/src/lib.rs
@@ -5,12 +5,15 @@
 extern crate libc;
 
 // Module declaration order is load-bearing. cbindgen emits functions in source
-// walk order (module declaration order, then top-to-bottom within each module),
-// and `ffi/zerobus.h` must stay byte-identical for the Go/Java ABI. The exported
-// functions must concatenate as: arrow fns + `zerobus_free_headers` (arrow) →
-// builder fns → sdk fns → stream fns ending with `zerobus_free_error_message`
-// and `zerobus_get_default_config` → proto fns. `common` declares the shared
-// types and has no exported functions.
+// walk order (module declaration order, then top-to-bottom within each module).
+// The generated `ffi/zerobus.h` declares the exported functions in that order;
+// C symbols link by name, so the order is not an ABI contract, but keeping it
+// stable minimises header churn. The exported functions concatenate as: the
+// shared header helpers from `common` (`zerobus_alloc_header_array`,
+// `zerobus_alloc_cstring`, `zerobus_free_headers`) → arrow fns → builder fns →
+// sdk fns → stream fns ending with `zerobus_free_error_message` and
+// `zerobus_get_default_config` → proto fns. `common` also declares the shared
+// types and the header helpers listed above.
 //
 // A comment between each declaration breaks rustfmt's contiguous-`mod` grouping,
 // so it cannot alphabetize them and reorder the generated header.
